try to modified reading.c and resize the app

+++ Slave/header/crc.h
@@ -7,9 +7,9 @@
 #define TOPBIT (1 << (WIDTH-1))
 
 
-int addCRC(int packetCrc, unsigned char countedCrc);
+uint16_t addCRC(uint16_t packetCrc, unsigned char countedCrc);
 
-int compareCRC(int crc1, int crc2);
+uint8_t compareCRC(uint16_t crc1, uint16_t crc2);
